Move all code to src parent folder
