+++ src/strategem.py
