+++ src/strategems.json
